gradle.properties

+++ jar/src/main/java/apc/jar/Android.kt
@@ -36,10 +36,10 @@ fun lint(project: String = "") {
     if (Files.isDirectory(root)) {
         root("$root/gradlew.bat", "lintDebug")
         Files.list(root).filter { Files.isDirectory(it) && rootIgnore.none(it::matchName) }.forEach {
-            val lint = it + "build/reports/lint-results-debug.xml"
-            if (Files.isRegularFile(lint)) {
-                println("\nParsing: $lint\n")
-                UnusedResources(lint)
+            val xml = it + "build/reports/lint-results-debug.xml"
+            if (Files.isRegularFile(xml)) {
+                println("\nParsing: $xml\n")
+                UnusedResources(xml)
             }
         }
     }
@@ -50,16 +50,14 @@ fun zipProject(project: String, copy: Boolean = false) {
     if (Files.isDirectory(source)) {
         val target = source.resolveSibling("${source.fileName}_mini")
         copyProject(source, target)
+        if (copy) println("Copied: ${target.toAbsolutePath()}")
 
         val zip = source.resolveSibling("${target.fileName}.zip")
         Files.deleteIfExists(zip)
         val targetFile = target.toFile()
         ZipFile(zip.toFile()).addFolder(targetFile, ZipParameters())
 
-        if (copy)
-            println("Copied: ${target.toAbsolutePath()}")
-        else
-            targetFile.deleteRecursively()
+        if (!copy) targetFile.deleteRecursively()
     }
 }
 
@@ -67,7 +65,7 @@ fun copyProject(source: Path, target: Path) {
     target.toFile().deleteRecursively()
 
     source copyTo target
-    val vcs = matcherList(".git*", ".svn")
+    val vcs = matcherList(".git*", ".svn", "gradle.properties")
     val moduleIgnore = matcherList(".*", "build", "*.iml")
     Files.list(source).parallel().filter { vcs.any(it::matchName) || rootIgnore.none(it::matchName) }.forEach {
         val targetModule = target + it.fileName
